Change the end determination policy
The block end was determined as the address do the last instruction in the
block.
Since now I want to use blocks to feed the emulator it makes more sense
for the end address to be the address of the first instruction of the next
block. This is particularly true for CISC architectures where instruction
size is variable length.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>