plugin: Switch to rizin
This patch switches the plugin to be compatible with rizin instead of
radare2.

The plugin registration mechanics are a bit different than in radare2,
i.e. command registration is done in an initialization function. There's
no need to check the command line anymore, so that code has been
removed. The structures in rizin are prefixed with 'Rz' instead of 'R'.

There are many changes in the makefile. The most important one is
switching to pkg-config for providing the include/linker flags.
This is an example how rizin should be compiled:

 $ meson --buildtype=release --prefix=~/.local -Duse_sys_capstone=enabled build
 $ cd build
 $ ninja install

Rizin can use 'capstone' either its 'subprojects' directory or
the system one. If you choose former however, it will be linked
statically. This is problematic because 'sysno' plugin wants to link
with 'capstone' as well. To make this work we would need to have
include/linker instructions in our makefile to help 'sysno' build
process find the version of 'capstone' from the 'subprojects' directory.
If you have 'capstone-devel' installed in OS-level, 'sysno' will use the
OS 'capstone' in build time and statically-compiled-in-to-rizin version
in runtime, which will make it crash in many magical ways...
It's much easier and safer to make both 'rizin' and 'sysno' use the
OS-provided version of 'capstone'.

You may need to define 'PKG_CONFIG_PATH' variable when building 'sysno'
plugin to help 'pkg-config' find rizin's '.pc' files in the
'~/.local' prefix:

$ export PKG_CONFIG_PATH="/home/myuser/.local/lib64/pkgconfig"

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>